added new facilityDropDown module
